refactor: :recycle: refactor logging module
